+++ src/AddTodoForm.js
@@ -1,0 +1,12 @@
+import React from 'react';
+function AddTodoForm() {
+return (
+    <form>
+        <lable htmlFor="todoTitle">Title</lable>
+        <input id="todoTitle"></input>
+        <button type="submit">Add</button>
+    </form>
+)
+}
+export default AddTodoForm;
+
