+++ src/App.js
@@ -1,32 +1,16 @@
 import React from 'react';
 
-const todoList = [
-  {
-    id: 1,
-    title: "Read the book",
-  },
-  {
-    id: 2,
-    title: "Watch video",
-  },
-  {
-    id: 3,
-    title: "Complete assignment",
-  }
-]
+import TodoList from './TodoList';
+import AddTodoForm from './AddTodoForm';
+
 
 function App() {
   return (
     <div>
       <h1>Todo List</h1> 
-      <ul>
-        {todoList.map(function(item) {
-          return <li key={item.id}>
-            <span>{item.title}</span>
-            </li>
-        })}
+     <TodoList />
+     <AddTodoForm/>
 
-      </ul>
 
     </div>
   );
